requirements.txt
flask==2.3.3
werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.26.0
python-multipart==0.0.5
pydub==0.25.1